+++ CHANGELOG.md
@@ -18,6 +18,7 @@ MSC's public documentation uses [Semantic Versioning](https://semver.org/) for i
 * improvements:
     * Add legend switch on interactive maps
     * Add scale line on interactive maps
+    * Add a label to distinguish between observation/extrapolation on weather radar interactive maps
 * bug fixes:
     * None